Bring together everything related to line parsing. And rename the class, since it now does nothing related to blocks, whatsoever.
