+++ src/Core/Gibberish/Gibberish.csproj
@@ -102,9 +102,8 @@
     <Compile Include="AST\ParseError.cs" />
     <Compile Include="AST\_1_Bare\CommentDefinitionBlock.cs" />
     <Compile Include="Parsing\ParsingExtensions.cs" />
-    <Compile Include="Parsing\RecognizeBlocks.cs" />
     <Compile Include="AST\_1_Bare\UnknownPrelude.cs" />
-    <Compile Include="Parsing\RecognizeBlocksParser.cs" />
+    <Compile Include="Parsing\RecognizeLines.cs" />
     <Compile Include="Parsing\UnfixableError.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AST\Statement.cs" />
